# api/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tomas API",
    "version": "2.0.0",
    "description": "\n        Personal and Home finance projection and expense tracking application API.\n        \n        ## Authentication\n        Most endpoints require JWT authentication. Include the token in the Authorization header:\n        ```\n        Authorization: Bearer <your-jwt-token>\n        ```\n        \n        ## Parameter Naming Convention\n        All API endpoints use `snake_case` for query parameters and request body fields.\n        \n        ## Breaking Changes (v2.0.0)\n        - Removed legacy `/transactions?book_id=...` endpoint - use `/books/{id}/transactions` instead\n        - All parameters now use `snake_case` (e.g., `account_id`, `start_date`, `end_date`, `up_to_date`)\n      ",
    "contact": {
      "name": "Tomas API Support",
      "url": "https://github.com/levhita/tomas"
    },
    "license": {
      "name": "GPL-3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.tomas.example.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "admin"
          },
          "superadmin": {
            "type": "boolean",
            "example": true
          },
          "active": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "account_id": {
            "type": "integer",
            "example": 456
          },
          "category_id": {
            "type": "integer",
            "nullable": true,
            "example": 789
          },
          "amount": {
            "type": "number",
            "format": "float",
            "example": -50,
            "description": "Negative for expenses, positive for income"
          },
          "description": {
            "type": "string",
            "example": "Grocery shopping"
          },
          "note": {
            "type": "string",
            "nullable": true,
            "example": "Weekly groceries"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2024-01-15"
          },
          "exercised": {
            "type": "boolean",
            "example": true,
            "description": "Whether the transaction has been executed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
      "TransactionInput": {
        "type": "object",
        "required": [
          "account_id",
          "amount",
          "description",
          "date"
        ],
        "properties": {
          "account_id": {
            "type": "integer",
            "example": 456
          },
          "category_id": {
            "type": "integer",
            "nullable": true,
            "example": 789
          },
          "amount": {
            "type": "number",
            "format": "float",
            "example": -50,
            "description": "Negative for expenses, positive for income"
          },
          "description": {
            "type": "string",
            "example": "Grocery shopping"
          },
          "note": {
            "type": "string",
            "nullable": true,
            "example": "Weekly groceries"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2024-01-15"
          },
          "exercised": {
            "type": "boolean",
            "default": true,
            "example": true
          }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "book_id": {
            "type": "integer",
            "example": 123
          },
          "parent_category_id": {
            "type": "integer",
            "nullable": true,
            "example": null
          },
          "name": {
            "type": "string",
            "example": "Food & Dining"
          },
          "note": {
            "type": "string",
            "nullable": true,
            "example": "Restaurant and grocery expenses"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "expense"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        }
      },
      "CategoryInput": {
        "type": "object",
        "required": [
          "name",
          "book_id"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Food & Dining"
          },
          "note": {
            "type": "string",
            "nullable": true,
            "example": "Restaurant and grocery expenses"
          },
          "parent_category_id": {
            "type": "integer",
            "nullable": true,
            "example": null,
            "description": "Optional parent category ID for hierarchical categories"
          },
          "book_id": {
            "type": "integer",
            "example": 123
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "default": "expense",
            "example": "expense",
            "description": "Category type (ignored if parent_category_id is provided)"
          }
        }
      },
...
      "Team": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Family Budget Team"
          },
          "note": {
            "type": "string",
            "nullable": true,
            "example": "Team for managing family finances"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": null
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "collaborator",
              "viewer"
            ],
            "example": "admin",
            "description": "User role in the team (included when listing user teams)"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Resource not found"
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "error",
              "message": "Invalid request parameters",
              "code": "BAD_REQUEST"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "error",
              "message": "Authentication required",
              "code": "UNAUTHORIZED"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Access forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "error",
              "message": "Access forbidden",
              "code": "FORBIDDEN"
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/categories": {
      "post": {
        "summary": "Create a new category",
        "description": "Create a new category within a book. Supports hierarchical categories with parent-child relationships.\n\n**Hierarchy constraints:**\n- If parent_category_id is provided, that category must not have a parent itself\n- Parent category must belong to the same book\n- Child categories automatically inherit parent type (type parameter is ignored)\n- Categories can only be nested two levels deep\n",
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Failed to create category",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/teams": {
      "get": {
        "summary": "List all teams accessible to the current user",
        "description": "Retrieve all teams that the authenticated user is a member of, including their role in each team.",
        "tags": [
          "Teams"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of teams the user has access to",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Team"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "description": "Failed to fetch teams",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/transactions": {
      "post": {
        "summary": "Create a new transaction",
        "description": "Create a new transaction within an account. Requires write access to the account's book.",
        "tags": [
          "Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Failed to create transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "summary": "Get list of all users with team statistics",
        "description": "Retrieve a paginated list of all users with search and filtering capabilities. Admin only endpoint.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of users per page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search query for username"
          },
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "superadmin",
                "regular"
              ]
            },
            "description": "Filter by user role"
          }
        ],
        "responses": {
          "200": {
            "description": "List of users with pagination info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/User"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "team_count": {
                                "type": "integer",
                                "description": "Number of teams user belongs to"
                              }
                            }
                          }
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Failed to fetch users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}